requirements.txt
tensorflow==2.16.1
Flask==3.0.0
numpy==1.24.3
pillow==10.0.0
gunicorn==21.2.0
h5py==3.11.0
